+++ module-service/src/main/java/com/mjc/school/service/impl/NewsService.java
@@ -2,17 +2,16 @@ package com.mjc.school.service.impl;
 
 
 import com.mjc.school.repository.BaseRepository;
+import com.mjc.school.repository.annotation.OnDelete;
 import com.mjc.school.repository.model.impl.NewsModel;
 import com.mjc.school.service.BaseService;
-import com.mjc.school.service.annotation.ValidateNewsId;
 import com.mjc.school.service.annotation.ValidateNewsDto;
+import com.mjc.school.service.annotation.ValidateNewsId;
 import com.mjc.school.service.dto.NewsDtoRequest;
 import com.mjc.school.service.dto.NewsDtoResponse;
 import com.mjc.school.service.exception.NotFoundException;
 import com.mjc.school.service.exception.ServiceErrorCodeMessage;
 import com.mjc.school.service.mapper.NewsModelMapper;
-import com.mjc.school.service.validator.Validator;
-import org.mapstruct.factory.Mappers;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -22,42 +21,34 @@ import java.util.List;
 @Service
 public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse, Long> {
 
-
-//    private NewsRepository baseRepository;
-
     private final BaseRepository<NewsModel, Long> baseRepository;
-    private final Validator newsValidator;
 
-    private NewsModelMapper mapper = (NewsModelMapper)Mappers.getMapper((Class)NewsModelMapper.class);
+    private NewsModelMapper mapper;
 
-    public NewsService(BaseRepository<NewsModel, Long> baseRepository, Validator newsValidator) {
+    public NewsService(BaseRepository<NewsModel, Long> baseRepository, NewsModelMapper mapper) {
         this.baseRepository = baseRepository;
-        this.newsValidator = newsValidator;
+        this.mapper = mapper;
 //        this.mapper = (NewsModelMapper)Mappers.getMapper((Class)NewsModelMapper.class);
     }
 
     @Override
     public List<NewsDtoResponse> readAll() {
-        return (List<NewsDtoResponse>)this.mapper.modelListToDtoList(this.baseRepository.readAll());
+        return (List<NewsDtoResponse>) this.mapper.modelListToDtoList(this.baseRepository.readAll());
     }
 
     @Override
     @ValidateNewsId
-    public NewsDtoResponse readById(Long newsId) {
-//        this.newsValidator.validateNewsId(newsId);
-//        if (this.baseRepository.existById(newsId)) {
-//            NewsModel newsModel = this.baseRepository.readById(newsId).get();
-//            return this.mapper.modelToDto(newsModel);
-//        }
-//        throw new NotFoundException(String.format(ServiceErrorCodeMessage.NEWS_ID_DOES_NOT_EXIST.getCodeMsg(), newsId));
-        NewsModel newsModel = this.baseRepository.readById(newsId).get();
-        return this.mapper.modelToDto(newsModel);
+    public NewsDtoResponse readById(Long id) {
+        if (this.baseRepository.existById(id)) {
+            NewsModel newsModel = this.baseRepository.readById(id).get();
+            return this.mapper.modelToDto(newsModel);
+        }
+        throw new NotFoundException(String.format(ServiceErrorCodeMessage.NEWS_ID_DOES_NOT_EXIST.getCodeMsg(), id));
     }
 
     @Override
     @ValidateNewsDto
     public NewsDtoResponse create(NewsDtoRequest dtoRequest) {
-//        this.newsValidator.validateNewsDto(dtoRequest);
         NewsModel model = this.mapper.dtoToModel(dtoRequest);
         LocalDateTime date = LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS);
         model.setCreateDate(date);
@@ -71,33 +62,25 @@ public class NewsService implements BaseService<NewsDtoRequest, NewsDtoResponse,
     @ValidateNewsId
     @ValidateNewsDto
     public NewsDtoResponse update(NewsDtoRequest dtoRequest) {
-//        this.newsValidator.validateNewsId(dtoRequest.id());
-//        this.newsValidator.validateNewsDto(dtoRequest);
-//        if (this.baseRepository.existById(dtoRequest.id())) {
-//            NewsModel model = this.mapper.dtoToModel(dtoRequest);
-//            LocalDateTime date = LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS);
-//            model.setLastUpdatedDate(date);
-//            NewsModel newsModel = this.baseRepository.update(model);
-//            return this.mapper.modelToDto(newsModel);
-//        }
-//        throw new NotFoundException(String.format(ServiceErrorCodeMessage.NEWS_ID_DOES_NOT_EXIST.getCodeMsg(), dtoRequest.id()));
-        NewsModel model = this.mapper.dtoToModel(dtoRequest);
-        LocalDateTime date = LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS);
-        model.setLastUpdatedDate(date);
-        NewsModel newsModel = this.baseRepository.update(model);
-        return this.mapper.modelToDto(newsModel);
+        if (this.baseRepository.existById(dtoRequest.id())) {
+            NewsModel model = this.mapper.dtoToModel(dtoRequest);
+            LocalDateTime date = LocalDateTime.now().truncatedTo(ChronoUnit.SECONDS);
+            model.setLastUpdatedDate(date);
+            NewsModel newsModel = this.baseRepository.update(model);
+            return this.mapper.modelToDto(newsModel);
+        }
+        throw new NotFoundException(String.format(ServiceErrorCodeMessage.NEWS_ID_DOES_NOT_EXIST.getCodeMsg(), dtoRequest.id()));
+
     }
 
     @Override
     @ValidateNewsId
+    @OnDelete
     public boolean deleteById(Long newsId) {
-//        this.newsValidator.validateNewsId(newsId);
-//        if (this.baseRepository.existById(newsId)) {
-//            return this.baseRepository.deleteById(newsId);
-//        }
-//        throw new NotFoundException(String.format(ServiceErrorCodeMessage.NEWS_ID_DOES_NOT_EXIST.getCodeMsg(), newsId));
-        return this.baseRepository.deleteById(newsId);
-
+        if (this.baseRepository.existById(newsId)) {
+            return this.baseRepository.deleteById(newsId);
+        }
+        throw new NotFoundException(String.format(ServiceErrorCodeMessage.NEWS_ID_DOES_NOT_EXIST.getCodeMsg(), newsId));
     }
 
 }
